backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pdfplumber==0.10.3
python-docx==1.1.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
pydantic==2.8.2
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
gunicorn==21.2.0